+++ Node.js/Personal-Projects/TODO-list_web-app/app/server.js
@@ -17,12 +17,11 @@ const app = express();
 app.use(logger('dev'));
 app.use(helmet());
 app.use(compression());
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded( {extended: true} ));
 
 // Serve Public FrontPage and static files:
 app.use('/', express.static(path.join(__dirname, './public')));
-
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded( {extended: true} ));
 
 mongoDB.MongoClient.connect(URI, (err, database) => {
   if (err)
@@ -33,18 +32,14 @@ mongoDB.MongoClient.connect(URI, (err, database) => {
 
   // Use the MongoDB present connection across the router:
   app.use((req, res, next) => {
-    res.locals.db = DB;
+    req.db = DB;
     next();
   });
 
-  auth(app);
+  auth(app, DB);
 
-  // TODO: Change to not static + check for authentication.
-  // Serve WebApp Page (Dashboard) and static files:
-  app.use('/dashboard', express.static(path.join(__dirname, './app-pages')));
-
-  // API:
-  app.use('/api', routes);
+  // ROUTES:
+  app.use('/', routes);
 
   const PORT = process.env.PORT || 3000;
   app.listen(PORT, () => {